- library fixes:
<b>getgrgid, getgrnam, getpwnam</b>
<blockquote>
These functions from grp.h in the Posix and Unix libraries should be
declared to return /*@observer@*/ references.  Reported by Enrico
Scholz.  Fixed in CVS (2002-07-08), will be in 3.0.1.7.
</blockquote>

<b>sys/resource.h</b>
<blockquote>
The sys/resource.h segment of the Unix library was not updated to
SUSv3.  The type of rlim_t, struct rlimit and struct rusage have been
updated.  Reported by Enrico Scholz.  Fixed in CVS (2002-07-08), will be
in 3.0.1.7.
</blockquote>

<b>Standard library missing snprintf</b>
<blockquote>
snprintf is in the unix library, but not the standard library.  It was
added to ISO C99, so should be in the standard library.  Reported by
Olivier L'Heureux.  Fixed in CVS (2002-07-08), will be in 3.0.1.7.
</blockquote>

- parsing support for restrict pointers

- showloadloc to control printing line number for load files