 Put back previously #if 0 'ed code.
I had to put this code back the case that needs it is when splint is given an absolute path name of a file outside of the current directory and the subdirectories below the current directory. e.g. cd /home/; splint /tmp/prog.c

NOTE: I added this code back previously in another branch.  I don't know if my changes were lost in CVS of if another developer removed them intentionally.